refactor: moved all types to a single file
